+++ libmesh/include/face_inf_quad6.h
@@ -1,4 +1,4 @@
-// $Id: face_inf_quad6.h,v 1.15 2003-08-07 19:25:31 ddreyer Exp $
+// $Id: face_inf_quad6.h,v 1.16 2003-08-18 14:44:52 ddreyer Exp $
 
 // The Next Great Finite Element Library.
 // Copyright (C) 2002  Benjamin S. Kirk, John W. Peterson
@@ -107,8 +107,8 @@ public:
    * that defines the \f$ n^{th} \f$ second-order node.
    * Note that \p n is counted as depicted above, \f$ 4 \le n < 6 \f$.
    */
-  unsigned int second_order_adjacent_vertex (const unsigned int n,
-					     const unsigned int v) const;
+  unsigned short int second_order_adjacent_vertex (const unsigned int n,
+						   const unsigned int v) const;
   
   
   
@@ -140,7 +140,7 @@ private:
    * Matrix that tells which vertices define the location
    * of mid-side (or second-order) nodes
    */
-  static const unsigned int _second_order_adjacent_vertices[2][2];
+  static const unsigned short int _second_order_adjacent_vertices[2][2];
     
 };
