+++ libmesh/include/face_tri6.h
@@ -1,4 +1,4 @@
-// $Id: face_tri6.h,v 1.13 2003-08-07 19:25:31 ddreyer Exp $
+// $Id: face_tri6.h,v 1.14 2003-08-18 14:44:52 ddreyer Exp $
 
 // The Next Great Finite Element Library.
 // Copyright (C) 2002  Benjamin S. Kirk, John W. Peterson
@@ -116,8 +116,8 @@ public:
    * that defines the \f$ n^{th} \f$ second-order node.
    * Note that \p n is counted as depicted above, \f$ 3 \le n < 6 \f$.
    */
-  unsigned int second_order_adjacent_vertex (const unsigned int n,
-					     const unsigned int v) const;
+  unsigned short int second_order_adjacent_vertex (const unsigned int n,
+						   const unsigned int v) const;
  
   
 private:
@@ -148,7 +148,7 @@ private:
    * Matrix that tells which vertices define the location
    * of mid-side (or second-order) nodes
    */
-  static const unsigned int _second_order_adjacent_vertices[3][2];
+  static const unsigned short int _second_order_adjacent_vertices[3][2];
     
 };
